moves attack network to attacker logic
