+++ insecure/adversary/attacker.go
@@ -7,6 +7,7 @@ import (
 	"sync"
 
 	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/hashicorp/go-multierror"
 	"github.com/rs/zerolog"
 	"google.golang.org/grpc"
 
@@ -22,20 +23,25 @@ const networkingProtocolTCP = "tcp"
 // Attacker implements the adversarial domain that is orchestrating an attack through corrupted nodes.
 type Attacker struct {
 	component.Component
-	address      string
-	server       *grpc.Server
-	logger       zerolog.Logger
-	orchestrator insecure.AttackOrchestrator
-	codec        network.Codec
-	cm           *component.ComponentManager
+	cm                 *component.ComponentManager
+	address            string
+	server             *grpc.Server
+	logger             zerolog.Logger
+	orchestrator       insecure.AttackOrchestrator
+	codec              network.Codec
+	corruptedIds       flow.IdentityList
+	corruptedNodes     map[flow.Identifier]insecure.CorruptedNodeConnection
+	corruptedConnector insecure.CorruptedNodeConnector
 }
 
 func NewAttacker(logger zerolog.Logger, address string, codec network.Codec, orchestrator insecure.AttackOrchestrator) (*Attacker, error) {
 	attacker := &Attacker{
-		orchestrator: orchestrator,
-		logger:       logger,
-		codec:        codec,
-		address:      address,
+		orchestrator:       orchestrator,
+		logger:             logger,
+		codec:              codec,
+		address:            address,
+		corruptedConnector: &CorruptedConnector{attackerAddress: address},
+		corruptedNodes:     make(map[flow.Identifier]insecure.CorruptedNodeConnection),
 	}
 
 	// setting lifecycle management module.
@@ -64,12 +70,19 @@ func NewAttacker(logger zerolog.Logger, address string, codec network.Codec, orc
 			ready()
 		}).
 		AddWorker(func(ctx irrecoverable.SignalerContext, ready component.ReadyFunc) {
-			attacker.start(ctx)
+			err := attacker.start(ctx)
+			if err != nil {
+				ctx.Throw(fmt.Errorf("could not start attacker: %w", err))
+			}
 
 			ready()
 
 			<-ctx.Done()
-			attacker.Stop()
+			err = attacker.stop()
+			if err != nil {
+				ctx.Throw(fmt.Errorf("could not stop attacker: %w", err))
+			}
+
 		}).Build()
 
 	attacker.Component = cm
@@ -79,13 +92,36 @@ func NewAttacker(logger zerolog.Logger, address string, codec network.Codec, orc
 }
 
 // start triggers the sub-modules of attacker.
-func (a *Attacker) start(ctx irrecoverable.SignalerContext) {
-	a.orchestrator.Start(ctx)
+func (a *Attacker) start(ctx irrecoverable.SignalerContext) error {
+	for _, corruptedId := range a.corruptedIds {
+		corruptibleClient, err := a.corruptedConnector.Connect(ctx, corruptedId.Address)
+		if err != nil {
+			return fmt.Errorf("could not establish corruptible client to node %x: %w", corruptedId.NodeID, err)
+		}
+		a.corruptedNodes[corruptedId.NodeID] = corruptibleClient
+	}
+
+	a.orchestrator.WithAttackNetwork(a)
+
+	return nil
 }
 
 // Stop stops the sub-modules of attacker.
-func (a *Attacker) Stop() {
+func (a *Attacker) stop() error {
+	// tears down connections to corruptible nodes.
+	var errors *multierror.Error
+	for _, connection := range a.corruptedNodes {
+		err := connection.CloseConnection()
+
+		if err != nil {
+			errors = multierror.Append(errors, err)
+		}
+	}
+
+	// tears down the attack server itself.
 	a.server.Stop()
+
+	return errors.ErrorOrNil()
 }
 
 // Observe implements the gRPC interface of attacker that is exposed to the corrupted conduits.
@@ -146,4 +182,89 @@ func (a *Attacker) processObservedMsg(message *insecure.Message) error {
 	}
 
 	return nil
+}
+
+// RpcUnicastOnChannel enforces unicast-dissemination on the specified channel through a corrupted node.
+func (a *Attacker) RpcUnicastOnChannel(corruptedId flow.Identifier, channel network.Channel, event interface{}, targetId flow.Identifier) error {
+	connection, ok := a.corruptedNodes[corruptedId]
+	if !ok {
+		return fmt.Errorf("no connection available for corrupted conduit factory to node %x: ", corruptedId)
+	}
+
+	msg, err := a.eventToMessage(corruptedId, event, channel, insecure.Protocol_UNICAST, 0, targetId)
+	if err != nil {
+		return fmt.Errorf("could not convert event to unicast message: %w", err)
+	}
+
+	err = connection.SendMessage(msg)
+	if err != nil {
+		return fmt.Errorf("could not sent unicast event to corrupted node: %w", err)
+	}
+
+	return nil
+}
+
+// RpcPublishOnChannel enforces a publish-dissemination on the specified channel through a corrupted node.
+func (a *Attacker) RpcPublishOnChannel(corruptedId flow.Identifier, channel network.Channel, event interface{},
+	targetIds ...flow.Identifier) error {
+	connection, ok := a.corruptedNodes[corruptedId]
+	if !ok {
+		return fmt.Errorf("no connection available for corrupted conduit factory to node %x: ", corruptedId)
+	}
+
+	msg, err := a.eventToMessage(corruptedId, event, channel, insecure.Protocol_PUBLISH, 0, targetIds...)
+	if err != nil {
+		return fmt.Errorf("could not convert event to publish message: %w", err)
+	}
+
+	err = connection.SendMessage(msg)
+	if err != nil {
+		return fmt.Errorf("could not sent publish event to corrupted node: %w", err)
+	}
+
+	return nil
+}
+
+// RpcMulticastOnChannel enforces a multicast-dissemination on the specified channel through a corrupted node.
+func (a *Attacker) RpcMulticastOnChannel(corruptedId flow.Identifier, channel network.Channel, event interface{}, num uint32,
+	targetIds ...flow.Identifier) error {
+	connection, ok := a.corruptedNodes[corruptedId]
+	if !ok {
+		return fmt.Errorf("no connection available for corrupted conduit factory to node %x: ", corruptedId)
+	}
+
+	msg, err := a.eventToMessage(corruptedId, event, channel, insecure.Protocol_MULTICAST, num, targetIds...)
+	if err != nil {
+		return fmt.Errorf("could not convert event to multicast message: %w", err)
+	}
+
+	err = connection.SendMessage(msg)
+	if err != nil {
+		return fmt.Errorf("could not sent multicast event to corrupted node: %w", err)
+	}
+
+	return nil
+}
+
+// eventToMessage converts the given application layer event to a protobuf message that is meant to be sent to the corrupted node.
+func (a *Attacker) eventToMessage(corruptedId flow.Identifier,
+	event interface{},
+	channel network.Channel,
+	protocol insecure.Protocol,
+	num uint32,
+	targetIds ...flow.Identifier) (*insecure.Message, error) {
+
+	payload, err := a.codec.Encode(event)
+	if err != nil {
+		return nil, fmt.Errorf("could not encode event: %w", err)
+	}
+
+	return &insecure.Message{
+		ChannelID: channel.String(),
+		OriginID:  corruptedId[:],
+		Targets:   num,
+		TargetIDs: flow.IdsToBytes(targetIds),
+		Payload:   payload,
+		Protocol:  protocol,
+	}, nil
 }
